requirements.txt
pandas
matplotlib
requests
Pillow
python-docx
